+++ packages/playground/ssr-react/server.js
@@ -9,24 +9,11 @@ async function createServer(
   root = process.cwd(),
   isProd = process.env.NODE_ENV === 'production'
 ) {
-  const toAbsolute = (p) => path.resolve(__dirname, p)
+  const resolve = (p) => path.resolve(__dirname, p)
 
   const indexProd = isProd
-    ? fs.readFileSync(toAbsolute('dist/client/index.html'), 'utf-8')
+    ? fs.readFileSync(resolve('dist/client/index.html'), 'utf-8')
     : ''
-
-  const manifest = isProd
-    ? // @ts-ignore
-      require('./dist/client/ssr-manifest.json')
-    : {}
-
-  function getIndexTemplate() {
-    if (isProd) {
-      return indexProd
-    }
-    // during dev, inject vite client + always read fresh index.html
-    return fs.readFileSync(toAbsolute('index.html'), 'utf-8')
-  }
 
   const app = express()
 
@@ -47,37 +34,42 @@ async function createServer(
   } else {
     app.use(require('compression')())
     app.use(
-      require('serve-static')(toAbsolute('dist/client'), {
+      require('serve-static')(resolve('dist/client'), {
         index: false
       })
     )
   }
 
-  app.use('*', async (req, res, next) => {
+  app.use('*', async (req, res) => {
     try {
       const url = req.originalUrl
-      let template = getIndexTemplate()
+
+      let template, render
       if (!isProd) {
+        // always read fresh template in dev
+        template = fs.readFileSync(resolve('index.html'), 'utf-8')
         template = await vite.transformIndexHtml(url, template)
+        render = (await vite.ssrLoadModule('/src/entry-server.jsx')).render
+      } else {
+        template = indexProd
+        render = require('./dist/server/entry-server.js').render
       }
 
-      const { render } = isProd
-        ? // @ts-ignore
-          require('./dist/server/entry-server.js')
-        : await vite.ssrLoadModule('/src/entry-server.ts')
+      const context = {}
+      const appHtml = render(url, context)
 
-      const [appHtml, preloadLinks] = await render(url, manifest)
+      if (context.url) {
+        // Somewhere a `<Redirect>` was rendered
+        return res.redirect(301, context.url)
+      }
 
-      const html = `
-      ${preloadLinks}
-      ${template.replace(`<!--ssr-outlet-->`, appHtml)}
-      `
+      const html = template.replace(`<!--app-html-->`, appHtml)
 
       res.status(200).set({ 'Content-Type': 'text/html' }).end(html)
     } catch (e) {
       !isProd && vite.ssrFixStacktrace(e)
       console.log(e.stack)
-      next(e)
+      res.status(500).end(e.stack)
     }
   })