+++ packages/playground/ssr-vue/src/App.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <router-link to="/">Home</router-link> | 
+    <router-link to="/about">About</router-link>
+    <router-view/>
+  </div>
+</template>
+
+<style>
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+</style>