+++ packages/playground/ssr/src/vue/App.vue
@@ -1,9 +1,0 @@
-<template>
-  <Async />
-</template>
-
-<script setup>
-import { defineAsyncComponent } from 'vue'
-
-const Async = defineAsyncComponent(() => import('./Async.vue'))
-</script>