+++ packages/playground/ssr/src/vue/index.ts
@@ -1,6 +1,0 @@
-import VueApp from './App.vue'
-import { createSSRApp } from 'vue'
-
-export function createVueApp() {
-  return createSSRApp(VueApp)
-}