+++ packages/vite/src/node/plugins/reporter.ts
@@ -26,7 +26,7 @@ export function buildReporterPlugin(config: ResolvedConfig): Plugin {
   const chunkLimit = config.build.chunkSizeWarningLimit
 
   async function getCompressedSize(code: string | Uint8Array): Promise<string> {
-    if (config.build.ssr) {
+    if (config.build.ssr || !config.build.brotliSize) {
       return ''
     }
     return ` / brotli: ${(
@@ -92,13 +92,15 @@ export function buildReporterPlugin(config: ResolvedConfig): Plugin {
     },
 
     buildEnd() {
-      if (tty) {
-        process.stdout.clearLine(0)
-        process.stdout.cursorTo(0)
+      if (shouldLogInfo) {
+        if (tty) {
+          process.stdout.clearLine(0)
+          process.stdout.cursorTo(0)
+        }
+        config.logger.info(
+          `${chalk.green(`✓`)} ${transformedCount} modules transformed.`
+        )
       }
-      config.logger.info(
-        `${chalk.green(`✓`)} ${transformedCount} modules transformed.`
-      )
     },
 
     renderStart() {
@@ -121,7 +123,7 @@ export function buildReporterPlugin(config: ResolvedConfig): Plugin {
     },
 
     generateBundle() {
-      if (tty) {
+      if (shouldLogInfo && tty) {
         process.stdout.clearLine(0)
         process.stdout.cursorTo(0)
       }